Moved grpc related python code to separate module.
